fixed gameOverViewController, moved a player into GameManager
