+++ Millionaire 1 team/GameManager.swift
@@ -6,8 +6,15 @@
 //
 
 import Foundation
+import AVFoundation
 
 final class GameManager {
+    
+    enum SoundType: String {
+        case chosenAnswer, win, lose, winGame, timerForResponse
+    }
+    
+    private var player: AVAudioPlayer?
     
     let questions = Question.getQuestions()
     
@@ -28,5 +35,23 @@ final class GameManager {
     func getCurrentQuestion() -> Question {
         let currentQuestion = levelQuestions[Int.random(in: 0..<levelQuestions.count)]
         return currentQuestion
+    }
+    
+    func playSound(type: SoundType) {
+        
+        guard let url = Bundle.main.url(forResource: type.rawValue, withExtension: "mp3") else { return }
+        
+        do {
+            try AVAudioSession.sharedInstance().setCategory(AVAudioSession.Category.playback)
+            try AVAudioSession.sharedInstance().setActive(true)
+            
+            player = try AVAudioPlayer(contentsOf: url)
+            
+            if let player = player {
+                player.play()
+            }
+        } catch let error {
+            print(error.localizedDescription)
+        }
     }
 }
